move mapgen into procgen
